Revert `Error` back to an unscoped enumeration (#5214)
Making it a scoped enumeration, coupled with the fact that `firebase::Future::error()` returns a plain `int`, imposes a lot of casting.

+++ Firestore/core/test/firebase/firestore/remote/serializer_test.cc
@@ -1591,8 +1591,7 @@ TEST_F(SerializerTest, DecodesListenResponseWithRemovedTargetChange) {
   change->add_target_ids(1);
   change->add_target_ids(2);
   change->set_resume_token("resume_token");
-  change->mutable_cause()->set_code(
-      static_cast<int32_t>(Error::kPermissionDenied));
+  change->mutable_cause()->set_code(Error::kPermissionDenied);
   change->mutable_cause()->set_message("Error message");
 
   SCOPED_TRACE("DecodesListenResponseWithRemovedTargetChange");
